==> final-project/1_final-project-NLP-Twitter_API_query.ipynb <==
# Twitter API query

Ressources : https://stackoverflow.com/questions/44116194/import-a-function-from-another-ipynb-file

import pandas as pd
from datetime import datetime
import re
import GetOldTweets3 as got #https://github.com/Jefferson-Henrique/GetOldTweets-python

start_time4 = datetime.now()

tweetCriteria = got.manager.TweetCriteria().setQuerySearch('#reformedesretraites')\
                                           .setLang("french")\
                                           .setSince("2019-12-05")\
                                           .setUntil("2019-12-31")\
                                           .setMaxTweets(500000)
tweet = got.manager.TweetManager.getTweets(tweetCriteria)

text = [x.text for x in tweet]

date = [x.date for x in tweet]

greve_twitter = pd.DataFrame(columns=['polarity','date','tweet'])

greve_twitter['tweet']=text
greve_twitter['date']=date

greve_twitter.head()

from ipynb.fs.full.my_functions import clean_up
greve_twitter['tweet_processed'] = greve_twitter['tweet'].astype(str).apply(lambda s: clean_up(s))

from ipynb.fs.full.my_functions import stem_spacy
greve_twitter['tweet_processed2'] = stem_spacy(greve_twitter['tweet_processed'])

greve_twitter['tweet_processed2'] = greve_twitter['tweet_processed2'].apply(lambda x: ' '.join(x))
greve_twitter.head()

greve_twitter.to_csv('greve_twitter_final.csv')

end_time4 = datetime.now()
print('Duration: {}'.format(end_time4 - start_time4))
