==> final-project/2_final-project-NLP_model1.ipynb <==
# <center>NLP with french tweets dataset</center>

Origin of train dataset  : https://github.com/gamebusterz/French-Sentiment-Analysis-Dataset (sample of 300.000, origin = './nlp/tweets.csv' (concatenate=cat x* > tweets.csv))
How to :
    1. train from French sentiment analysis dataset
    2. cleaning : cleaning, token, stop-words then split in train/test
    3. Count vectorization
    4. NLP classification models : xgboost, logreg, naive bayes
    5. Scores and time evaluation


from datetime import datetime
start_time0 = datetime.now()
import spacy
import numpy as np
import re
import nltk
import pandas as pd
pd.set_option('display.max_colwidth', None)
import pickle
import xgboost as xgb
from ipynb.fs.full.my_functions import clean_up, stem_spacy
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
from nltk.stem.snowball import FrenchStemmer
from nltk import NaiveBayesClassifier
from sklearn.svm import SVC
from sklearn.model_selection import train_test_split,  GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.linear_model import LogisticRegression 
from sklearn.metrics import accuracy_score, classification_report
from xgboost import XGBClassifier

## Step 1 :  ML model construction 

train= pd.read_csv('tweets.csv',index_col=[0]).reset_index(drop=True)
train.head()

train = train.reset_index(drop=True)

train.head()

train.columns


train = train.dropna()
train['polarity'] = train['polarity'].str.replace("0","negatif").str.replace("4","positif")
index_zero = train[(train['polarity'] != 'positif') & (train['polarity']!= 'negatif')].index
train.drop(index_zero, inplace=True)

train['polarity'].value_counts()
#.plot(kind='bar')

New column with clean tweet

from ipynb.fs.full.my_functions import clean_up
from ipynb.fs.full.my_functions import stem_spacy
train.statutnull = train.statutnull.astype(str)
train['original'] = train['statutnull']
train.statutnull = train.statutnull.apply(lambda s: clean_up(s))
train['tweet_stem'] = stem_spacy(train.statutnull)
train.rename(columns={'statutnull':'tweet'},inplace=True)
train.head()

train.shape

train.reset_index(drop=True).head()

New column for XGboost, SVM and Logistic regression and count vectorizer

train['tweet_stem_join'] = train.tweet_stem.apply(lambda x: ' '.join(x)).astype(str)
train.head()

count_vectorizer = CountVectorizer(analyzer='word') 
cv = count_vectorizer.fit(train['tweet_stem_join'])
pickle.dump(cv, open("cv1.pkl", "wb"))

cv_ = count_vectorizer.fit_transform(train['tweet_stem_join'])
cv_.shape

cv1 = pickle.load(open("cv1.pkl", 'rb'))
cv1_new = CountVectorizer(vocabulary = cv1.vocabulary_)
X_cv1 = cv1_new.transform(train['tweet_stem_join'])
X_cv1.shape

OR vectorization with tfidf module
from sklearn.feature_extraction.text import TfidfVectorizer  
tfidfconverter = TfidfVectorizer(max_features=2000, min_df=5, max_df=0.7, stop_words=stopwords.words('french'))  
X = tfidfconverter.fit_transform(train['tweet_stem_join']).toarray()

X_train,X_test,y_train,y_test = train_test_split(X,train['polarity'] , test_size=.2,
                                                 stratify=train['polarity'], random_state=42)
# Lower score with tfidf because the corpus is weak (traduction)
# https://datascience.stackexchange.com/questions/49047/will-a-count-vectorizer-ever-perform-slightly-better-than-tf-idf

end_time0 = datetime.now()
print('Time import and clean: {}'.format(end_time0 - start_time0))

## SVM

start_time_svc = datetime.now()
X_train_svc,X_test_svc,y_train_svc,y_test_svc = train_test_split(cv_,train['polarity'],test_size=.2,
                                                 stratify=train['polarity'], random_state=42)

grid = {'C':[1,10,100,1000],'gamma':[1,0.1,0.001,0.0001], 'kernel':['linear','rbf']}
svc = SVC()
svc_wv=GridSearchCV(svc,grid,cv=10)
svc_wv.fit(X_train_svc,y_train_svc)
print("tuned hpyerparameters :(best parameters) ",svc_wv.best_params_)
print("accuracy :",svc_wv.best_score_)

svc = SVC()
svc.fit(X_train_svc,y_train_svc)
prediction_svc = svc.predict(X_test_svc)
print(classification_report(prediction_svc,y_test_svc))
end_time_svc = datetime.now()
print('Temps modèle SVM : {}'.format(end_time_svc - start_time_svc))

## XGBoost model

start_time1 = datetime.now()
X_train_xg,X_test_xg,y_train_xg,y_test_xg = train_test_split(cv_,train['polarity'],test_size=.2,stratify=train['polarity'], random_state=42)

#grid search xgb
grid={"max_depth":[3,6,9],'min_child_weight':[4,5,6]}
xgbc = XGBClassifier(n_estimators=1000, nthread=6, objective= 'binary:logistic')
xgbc_cv=GridSearchCV(xgbc,grid,cv=10)
xgbc_cv.fit(X_train_xg,y_train_xg)
print("tuned hpyerparameters :(best parameters) ",xgbc_cv.best_params_)
print("accuracy :",xgbc_cv.best_score_)
#tuned hpyerparameters :(best parameters)  {'max_depth': 9, 'min_child_weight': 4}
#accuracy : 0.7395625

xgbc = XGBClassifier(max_depth=9, min_child_weight=4, n_estimators=1000, nthread= 3)
xgbc.fit(X_train_xg,y_train_xg)
prediction_xgb = xgbc.predict(X_test_xg)
print(classification_report(prediction_xgb,y_test_xg))
#0.742425 count vectorizer (max_depth=6, n_estimators=1000, nthread= 3)
#0.7415 tfidf
print(accuracy_score(prediction_xgb,y_test_xg))
#0.7531 300.000 sample 21/01
end_time1 = datetime.now()
print('Temps modèle XGBoost : {}'.format(end_time1 - start_time1))

## Logistic regression model

start_time2 = datetime.now()
X_train_logreg,X_test_logreg,y_train_logreg,y_test_logreg = train_test_split(cv_,train['polarity'] , test_size=.2,
                                                 stratify=train['polarity'], random_state=42)

#grid search logreg
grid={"C":np.logspace(-3,3,7), "penalty":["l1","l2"]}# l1 lasso l2 ridge
logreg=LogisticRegression()
cv_logreg=GridSearchCV(logreg,grid,cv=10)
cv_logreg.fit(X_train_logreg,y_train_logreg)

print("tuned hpyerparameters :(best parameters) ",cv_logreg.best_params_)
print("accuracy :",cv_logreg.best_score_)
#tuned hpyerparameters :(best parameters)  {'C': 0.1, 'penalty': 'l2'}
#accuracy : 0.7391625

logreg = LogisticRegression(C=0.1,penalty="l2")
logreg.fit(X_train_logreg,y_train_logreg)
prediction_logreg = logreg.predict(X_test_logreg)
print(classification_report(prediction_logreg,y_test_logreg))
#0.7628 count vectorizer
#0.74105 tfidf
print(accuracy_score(prediction_logreg,y_test_logreg))
end_time2 = datetime.now()
print('Time Logistic regression model: {}'.format(end_time2 - start_time2))

## Naive Bayes model

train2 = train.sample(100000)
start_time3 = datetime.now()
all_words = []
NUM_FEATURES = 5000
for index, value in train2.tweet.iteritems():
    if value not in all_words:
        all_words += value
top_features = [x[0] for x in nltk.FreqDist(all_words).most_common(NUM_FEATURES)]
df=pd.DataFrame.from_dict(nltk.FreqDist(all_words),orient='index')
df.sort_values(by=0, ascending=False).reset_index().head(10)
def build_features(words):
    features = {}
    for w in top_features:
        features[w] = (w in words)
    return features
featuresets = []
for index, row in train.iterrows():
    featuresets.append((build_features(row['tweet']), row['polarity']))
print(featuresets[:2])

train_set, test = train_test_split(featuresets, test_size=0.2)
classifier = NaiveBayesClassifier.train(train_set)
print(classifier.show_most_informative_features(n=10))
print(nltk.classify.accuracy(classifier, test))
# score 73,56% pour sample 100.000 avec stem nltk
# score 73,88% avec sample 100.000 avec stem spacy 

print(classification_report(classifier, test))
end_time3 = datetime.now()
print('Time Naive Bayes model: {}'.format(end_time3 - start_time3))

# Step 2 : applying final model on API twitter extraction 

greve_twitter=pd.read_csv('greve_twitter_final.csv',index_col=[0])
greve_twitter.head()

X_cv2 = cv1_new.transform(greve_twitter['tweet_processed2'])
greve_twitter['polarity']=logreg.predict(X_cv2)
greve_twitter.head()

greve_twitter.to_csv('greve_twitter_class.csv')
